RF: Split plugin and package managers for readability
